Agent: Move ExploiterResultData, FingerprintData, PingScanData under infection_monkey/dataclasses/
