+++ monkey/infection_monkey/dataclasses/__init__.py
@@ -1,1 +1,4 @@
+from .exploiter_result_data import ExploiterResultData
+from .fingerprint_data import FingerprintData
+from .ping_scan_data import PingScanData
 from .port_scan_data import PortScanData
