+++ monkey/infection_monkey/dataclasses/exploiter_result_data.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+from typing import Mapping, Optional
+
+
+@dataclass
+class ExploiterResultData:
+    exploitation_success: bool = False
+    propagation_success: bool = False
+    os: str = ""
+    info: Optional[Mapping] = None
+    error_message: str = ""
