+++ monkey/infection_monkey/dataclasses/fingerprint_data.py
@@ -1,0 +1,3 @@
+from collections import namedtuple
+
+FingerprintData = namedtuple("FingerprintData", ["os_type", "os_version", "services"])
