+++ monkey/infection_monkey/dataclasses/ping_scan_data.py
