+++ monkey/infection_monkey/exploit/HostExploiter.py
@@ -8,7 +8,7 @@ from common.agent_events import ExploitationEvent, PropagationEvent
 from common.event_queue import IAgentEventQueue
 from common.types import Event
 from common.utils.exceptions import FailedExploitationError
-from infection_monkey.i_puppet import ExploiterResultData
+from infection_monkey.dataclasses import ExploiterResultData
 from infection_monkey.model import TargetHost
 from infection_monkey.network import TCPPortSelector
 from infection_monkey.utils.ids import get_agent_id
