+++ monkey/infection_monkey/exploit/log4shell.py
@@ -10,6 +10,7 @@ from common.tags import (
     T1203_ATTACK_TECHNIQUE_TAG,
 )
 from common.utils import Timer
+from infection_monkey.dataclasses import ExploiterResultData
 from infection_monkey.exploit.log4shell_utils import (
     LINUX_EXPLOIT_TEMPLATE_PATH,
     WINDOWS_EXPLOIT_TEMPLATE_PATH,
@@ -21,7 +22,6 @@ from infection_monkey.exploit.log4shell_utils import (
 from infection_monkey.exploit.tools.helpers import get_agent_dst_path
 from infection_monkey.exploit.tools.http_tools import HTTPTools
 from infection_monkey.exploit.web_rce import WebRCE
-from infection_monkey.i_puppet.i_puppet import ExploiterResultData
 from infection_monkey.model import DROPPER_ARG, LOG4SHELL_LINUX_COMMAND, LOG4SHELL_WINDOWS_COMMAND
 from infection_monkey.network.tools import get_interface_to_target
 from infection_monkey.utils.commands import build_monkey_commandline
