+++ monkey/infection_monkey/exploit/mssqlexec.py
@@ -14,10 +14,10 @@ from common.tags import (
     T1210_ATTACK_TECHNIQUE_TAG,
 )
 from common.utils.exceptions import FailedExploitationError
+from infection_monkey.dataclasses import ExploiterResultData
 from infection_monkey.exploit.HostExploiter import HostExploiter
 from infection_monkey.exploit.tools.helpers import get_agent_dst_path
 from infection_monkey.exploit.tools.http_tools import HTTPTools
-from infection_monkey.i_puppet import ExploiterResultData
 from infection_monkey.model import DROPPER_ARG
 from infection_monkey.transport import LockedHTTPServer
 from infection_monkey.utils.brute_force import generate_identity_secret_pairs
