+++ monkey/infection_monkey/exploit/zerologon.py
@@ -23,13 +23,13 @@ from common.tags import (
     T1098_ATTACK_TECHNIQUE_TAG,
     T1210_ATTACK_TECHNIQUE_TAG,
 )
+from infection_monkey.dataclasses import ExploiterResultData
 from infection_monkey.exploit.HostExploiter import HostExploiter
 from infection_monkey.exploit.tools.wmi_tools import WmiTools
 from infection_monkey.exploit.zerologon_utils.dc_utils import connect_to_dc, get_dc_details
 from infection_monkey.exploit.zerologon_utils.dump_secrets import DumpSecrets
 from infection_monkey.exploit.zerologon_utils.options import OptionsForSecretsdump
 from infection_monkey.exploit.zerologon_utils.wmiexec import Wmiexec
-from infection_monkey.i_puppet import ExploiterResultData
 from infection_monkey.utils.capture_output import StdoutCapture
 from infection_monkey.utils.ids import get_agent_id
 from infection_monkey.utils.threading import interruptible_iter
