+++ monkey/infection_monkey/master/exploiter.py
@@ -7,7 +7,8 @@ from typing import Callable, Dict, Mapping, Sequence
 from common.agent_configuration.agent_sub_configurations import ExploitationConfiguration
 from common.types import Event
 from infection_monkey.custom_types import PropagationCredentials
-from infection_monkey.i_puppet import ExploiterResultData, IPuppet, RejectedRequestError
+from infection_monkey.dataclasses import ExploiterResultData
+from infection_monkey.i_puppet import IPuppet, RejectedRequestError
 from infection_monkey.model import TargetHost
 from infection_monkey.utils.threading import interruptible_iter, run_worker_threads
