+++ monkey/infection_monkey/master/propagator.py
@@ -11,8 +11,12 @@ from common.agent_configuration import (
     ScanTargetConfiguration,
 )
 from common.types import Event, NetworkPort, PortStatus
-from infection_monkey.dataclasses import PortScanData
-from infection_monkey.i_puppet import ExploiterResultData, FingerprintData, PingScanData
+from infection_monkey.dataclasses import (
+    ExploiterResultData,
+    FingerprintData,
+    PingScanData,
+    PortScanData,
+)
 from infection_monkey.model import TargetHost
 from infection_monkey.network import NetworkAddress
 from infection_monkey.network_scanning.scan_target_generator import compile_scan_target_list
