+++ monkey/infection_monkey/network_scanning/smb_fingerprinter.py
@@ -7,8 +7,8 @@ from odict import odict
 
 from common import OperatingSystem
 from common.types import PortStatus
-from infection_monkey.dataclasses import PortScanData
-from infection_monkey.i_puppet import FingerprintData, IFingerprinter, PingScanData
+from infection_monkey.dataclasses import FingerprintData, PingScanData, PortScanData
+from infection_monkey.i_puppet import IFingerprinter
 
 DISPLAY_NAME = "SMB"
 SMB_PORT = 445
