+++ monkey/infection_monkey/network_scanning/ssh_fingerprinter.py
@@ -2,8 +2,8 @@ import re
 from typing import Dict, Optional, Tuple
 
 from common import OperatingSystem
-from infection_monkey.dataclasses import PortScanData
-from infection_monkey.i_puppet import FingerprintData, IFingerprinter, PingScanData
+from infection_monkey.dataclasses import FingerprintData, PingScanData, PortScanData
+from infection_monkey.i_puppet import IFingerprinter
 
 SSH_REGEX = r"SSH-\d\.\d-OpenSSH"
 LINUX_DIST_SSH = ["ubuntu", "debian"]
